+++ files/root/index.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * Root index file for DDEV WP Local Setup.
+ *
+ * @package CassidyDC/DDEV-WP-Local-Setup
+ */
+
+// Load the WordPress index file.
+require_once __DIR__ . '/wordpress/index.php';