+++ CHANGELOG.md
@@ -1,4 +1,17 @@
-# Changelog for `ddev-local-wordpress-script`
+# Changelog for `ddev-wp-local-setup`
+
+## [2.3.0] - 2026-03-20
+
+### Added
+
+-   Added root `composer.json` file for copying.
+
+### Changed
+
+-   Moved project repo from `https://github.com/jacobcassidy` to `https://github.com/CassidyDC`.
+-   Refactored package structure.
+-   Renamed package from `ddev-local-wordpress-script` to `ddev-wp-local-setup`.
+-   Updated the install script to create project config files in the `/wp-content` directory, instead of the server root.
 
 ## [2.2.4] - 2025-10-19
 
@@ -12,7 +25,7 @@
 
 ## [2.2.3] - 2025-10-18
 
--   Replaced `ddev-wp-setup-script` text with `ddev-local-wordpress-script`
+-   Replaced `ddev-wp-setup-script` text with `ddev-wp-local-setup`
 
 ## [2.2.2] - 2025-10-18